Fix progress bar for non-square images, long scene loads.
I mixed up row and col in the names of the rendering code. Because I
made the same mistake when copying to the image buffer, it didn't
actually change the resulting image, only the counted number of
iterations for the progress bar.

I also fix the progress bar to only start counting time after building
the scene. Otherwise, it suddenly starts w/ several seconds on the
clock.