Move camera config into scene yaml.
We were already setting the view angle from the scene file, so it makes
sense to set other properties such as FOV there as well.

I also update the default FOV to 60 (which makes more sense for the
average image), and manaully set the FOV for the existing test scenes.